// angular/src/services/interfaces/models.ts
export interface Customer {
    id?: number;
    isActive?: boolean;
    isDeleted?: boolean;
    createdBy?: string | undefined;
    createdDate?: moment.Moment;
    modifiedBy?: string | undefined;
    modifiedDate?: moment.Moment | undefined;
    addressLine1?: string | undefined;
    addressLine2?: string | undefined;
    city?: string | undefined;
    state?: string | undefined;
    country?: string | undefined;
    zipCode?: string | undefined;
    phoneNumber?: string | undefined;
    faxNumber?: string | undefined;
    email?: string | undefined;
    latitude?: string | undefined;
    longitude?: string | undefined;
    accountNumber?: number;
    customerName?: string | undefined;
    customerTypeId?: number;
    customerType?: CustomerType;
    priceLevelId?: number | undefined;
    priceLevel?: PriceLevel;
    paymentTermId?: number | undefined;
    paymentTerm?: PaymentTerm;
    salesRepresentativeOutId?: number | undefined;
    salesRepresentativeOut?: SalesRepresentative;
    salesRepresentativeInId?: number | undefined;
    salesRepresentativeIn?: SalesRepresentative;
    defaultWarehouseId?: number | undefined;
    warehouse?: Warehouse;
    contactName?: string | undefined;
    taxRate?: number | undefined;
    overBalance?: number | undefined;
    creditLimit?: number | undefined;
    isBypassCreditLimit?: boolean;
    isTaxable?: boolean;
    isHoldAccount?: boolean;
    discount?: number | undefined;
    sellersPermit?: string | undefined;
    trialBalance?: number | undefined;
    accountsReceivableBalance?: number | undefined;
    crossStreet?: string | undefined;
    isWithShippingCost?: boolean;
    isWithHandlingCost?: boolean;
    isAllowViewCA?: boolean;
    isAllowViewVA?: boolean;
    isAllowViewWAG?: boolean;
    isOverrideCreditLimitRestriction?: boolean;
    shipperAccountNumber?: string | undefined;
    isApplyPFGShippingCharge?: boolean;
    isApplyPFGHandlingCharge?: boolean;
    isPFGShippingHandlingLevel?: boolean;
    freightPrepaidAmount?: number | undefined;
}
// ...
export interface Order {
    id?: number;
    isActive?: boolean;
    isDeleted?: boolean;
    createdBy?: string | undefined;
    createdDate?: moment.Moment;
    modifiedBy?: string | undefined;
    modifiedDate?: moment.Moment | undefined;
    customerId?: number;
    customer?: Customer;
    paymentTermId?: number | undefined;
    paymentTerm?: PaymentTerm;
    warehouseId?: number;
    warehouse?: Warehouse;
    orderNumber?: number | undefined;
    quoteNumber?: number | undefined;
    customerName?: string | undefined;
    accountNumber?: number;
    paymentTermName?: string | undefined;
    priceLevelId?: number;
    priceLevelName?: string | undefined;
    warehouseName?: string | undefined;
    orderDate?: moment.Moment;
    user?: string | undefined;
    orderStatusId?: number;
    orderStatusName?: string | undefined;
    orderedBy?: string | undefined;
    orderedByEmail?: string | undefined;
    orderedByPhoneNumber?: string | undefined;
    orderedByNotes?: string | undefined;
    deliveryType?: number;
    deliveryDate?: moment.Moment;
    deliveryRoute?: number | undefined;
    discount?: number | undefined;
    restockingFee?: number | undefined;
    restockingAmount?: number | undefined;
    taxRate?: number | undefined;
    totalTax?: number | undefined;
    subTotalAmount?: number | undefined;
    totalAmount?: number | undefined;
    isQuote?: boolean;
    isPrinted?: boolean;
    isCustomerOrder?: boolean;
    linkedInvoiceNumber?: string | undefined;
    orderDetails?: OrderDetail[] | undefined;
    email?: string | undefined;
    phoneNumber?: string | undefined;
    billAddress?: string | undefined;
    billCity?: string | undefined;
    billState?: string | undefined;
    billZipCode?: string | undefined;
    billZone?: string | undefined;
    billPhoneNumber?: string | undefined;
    billContactName?: string | undefined;
    shipAddressName?: string | undefined;
    shipAddress?: string | undefined;
    shipCity?: string | undefined;
    shipState?: string | undefined;
    shipZipCode?: string | undefined;
    shipZone?: string | undefined;
    shipPhoneNumber?: string | undefined;
    shipContactName?: string | undefined;
    isHoldInvoice?: boolean;
    invoiceNumber?: number | undefined;
    fwd?: number | undefined;
    originalInvoiceNumber?: number | undefined;
    orderType?: number | undefined;
    handlingType?: number | undefined;
    creditType?: number | undefined;
    singleInvoice?: number | undefined;
    purchaseOrderNumber?: string | undefined;
    type?: string | undefined;
    category?: string | undefined;
    carrier?: string | undefined;
    trackingNumber?: string | undefined;
    salesRepresentative1?: string | undefined;
    salesRepresentative2?: string | undefined;
    invoiceNotes?: string | undefined;
    particular?: string | undefined;
    paymentReference?: string | undefined;
    postedBy?: string | undefined;
    accountingPaidBy?: string | undefined;
    transactionDate?: moment.Moment | undefined;
    postDate?: moment.Moment | undefined;
    paymentDate?: moment.Moment | undefined;
    debit?: number | undefined;
    credit?: number | undefined;
    quantity?: number | undefined;
    quantityShipped?: number | undefined;
    currentCost?: number | undefined;
    amountPaid?: number | undefined;
    balance?: number | undefined;
    payment?: number | undefined;
    listPrice?: number | undefined;
    rgaType?: number;
    rgaReason?: number;
    rgaReasonNotes?: string | undefined;
}
// ...
export interface PaymentTerm {
    id?: number;
    isActive?: boolean;
    isDeleted?: boolean;
    createdBy?: string | undefined;
    createdDate?: moment.Moment;
    modifiedBy?: string | undefined;
    modifiedDate?: moment.Moment | undefined;
    termName?: string | undefined;
    description?: string | undefined;
    numberOfDays?: number;
    notes?: string | undefined;
}
// ...
export interface PurchaseOrder {
    id?: number;
    isActive?: boolean;
    isDeleted?: boolean;
    createdBy?: string | undefined;
    createdDate?: moment.Moment;
    modifiedBy?: string | undefined;
    modifiedDate?: moment.Moment | undefined;
    vendorId?: number;
    vendor?: Vendor;
    vendorName?: string | undefined;
    vendorCode?: string | undefined;
    vendorPO?: string | undefined;
    pfwbNumber?: string | undefined;
    statusId?: number;
    poStatus?: string | undefined;
    purchaseOrderDate?: moment.Moment;
    receivedDate?: moment.Moment | undefined;
    isPrinted?: boolean;
    totalQuantity?: number;
    totalAmount?: number;
    purchaseOrderDetails?: PurchaseOrderDetail[] | undefined;
}

export interface PurchaseOrderDetail {
    id?: number;
    isActive?: boolean;
    isDeleted?: boolean;
    createdBy?: string | undefined;
    createdDate?: moment.Moment;
    modifiedBy?: string | undefined;
    modifiedDate?: moment.Moment | undefined;
    purchaseOrderId?: number;
    purchaseOrder?: PurchaseOrder;
    orderDetailId?: number;
    orderDetail?: OrderDetail;
    orderId?: number;
    orderNumber?: number;
    orderQuantity?: number;
    shipZone?: string | undefined;
    partNumber?: string | undefined;
    vendorPartNumber?: string | undefined;
    vendorPrice?: number;
    sequence?: string | undefined;
    partDescription?: string | undefined;
    orderDate?: moment.Moment;
    deliveryDate?: moment.Moment;
    deliveryRoute?: number;
    statusId?: number;
    poDetailStatus?: string | undefined;
    deliveryMethod?: string | undefined;
    mainPartsLinkNumber?: string | undefined;
    purchaseOrderNumber?: string | undefined;
    customerName?: string | undefined;
    receivedDate?: moment.Moment | undefined;
}
// ...
export interface Role {
    id?: number;
    isActive?: boolean;
    isDeleted?: boolean;
    createdBy?: string | undefined;
    createdDate?: moment.Moment;
    modifiedBy?: string | undefined;
    modifiedDate?: moment.Moment | undefined;
    name?: string | undefined;
    code?: string | undefined;
    description?: string | undefined;
    sortOrder?: number;
    userTypeId?: number;
    userType?: UserType;
    notes?: string | undefined;
    rolePermissions?: RolePermission[] | undefined;
}
// ...
export interface RolePermission {
    id?: number;
    isActive?: boolean;
    isDeleted?: boolean;
    createdBy?: string | undefined;
    createdDate?: moment.Moment;
    modifiedBy?: string | undefined;
    modifiedDate?: moment.Moment | undefined;
    roleId?: number;
    role?: Role;
    moduleGroupId?: number;
    moduleGroup?: ModuleGroup;
    moduleId?: number;
    module?: Module;
    accessTypeId?: number;
    accessType?: AccessType;
}
// ...
export interface User {
    id?: number;
    isActive?: boolean;
    isDeleted?: boolean;
    createdBy?: string | undefined;
    createdDate?: moment.Moment;
    modifiedBy?: string | undefined;
    modifiedDate?: moment.Moment | undefined;
    roleId?: number;
    role?: Role;
    customerId?: number | undefined;
    customer?: Customer;
    locationId?: number | undefined;
    location?: Location;
    userName?: string | undefined;
    email?: string | undefined;
    isCustomerUser?: boolean;
    isActivated?: boolean;
    isChangePasswordOnLogin?: boolean;
    passwordHash?: string | undefined;
    passwordSalt?: string | undefined;
}
// ...
export interface UserType {
    id?: number;
    isActive?: boolean;
    isDeleted?: boolean;
    createdBy?: string | undefined;
    createdDate?: moment.Moment;
    modifiedBy?: string | undefined;
    modifiedDate?: moment.Moment | undefined;
    name?: string | undefined;
    code?: string | undefined;
    description?: string | undefined;
    sortOrder?: number;
}

export interface Vendor {
    id?: number;
    isActive?: boolean;
    isDeleted?: boolean;
    createdBy?: string | undefined;
    createdDate?: moment.Moment;
    modifiedBy?: string | undefined;
    modifiedDate?: moment.Moment | undefined;
    addressLine1?: string | undefined;
    addressLine2?: string | undefined;
    city?: string | undefined;
    state?: string | undefined;
    country?: string | undefined;
    zipCode?: string | undefined;
    phoneNumber?: string | undefined;
    faxNumber?: string | undefined;
    email?: string | undefined;
    latitude?: string | undefined;
    longitude?: string | undefined;
    vendorName?: string | undefined;
    vendorCode?: string | undefined;
    contactName?: string | undefined;
    isCAVendor?: boolean;
    isNVVendor?: boolean;
    caRank?: number;
    nvRank?: number;
    caPercentage?: number;
    nvPercentage?: number;
    cutoffTime?: moment.Moment;
}
// ...
export interface VendorOrderDTO {
    id?: number;
    isActive?: boolean;
    isDeleted?: boolean;
    createdBy?: string | undefined;
    createdDate?: moment.Moment;
    modifiedBy?: string | undefined;
    modifiedDate?: moment.Moment | undefined;
    orderId?: number;
    orderDetailId?: number;
    orderNumber?: number;
    purchaseOrderNumber?: string | undefined;
    mainPartsLinkNumber?: string | undefined;
    deliveryMethod?: string | undefined;
    sequence?: string | undefined;
    orderQuantity?: number;
    shipZone?: string | undefined;
    partNumber?: string | undefined;
    vendorPartNumber?: string | undefined;
    vendorPrice?: number;
    partDescription?: string | undefined;
    orderDate?: moment.Moment;
    poDetailStatus?: string | undefined;
    customerName?: string | undefined;
    deliveryDate?: moment.Moment;
    deliveryRoute?: number | undefined;
}

export interface Warehouse {
    id?: number;
    isActive?: boolean;
    isDeleted?: boolean;
    createdBy?: string | undefined;
    createdDate?: moment.Moment;
    modifiedBy?: string | undefined;
    modifiedDate?: moment.Moment | undefined;
    warehouseName?: string | undefined;
    description?: string | undefined;
    notes?: string | undefined;
}